cloud_storage_clients: split out multipart state and error helpers
An upcoming change will add a dependency on the client_pool to the
multipart upload code. Rather than making the *_clients dependent on
client_pool.h, this splits the multipart code out into their own .cc
files.

No functional changes.
