+++ src/v/cloud_storage_clients/BUILD
@@ -4,7 +4,9 @@ redpanda_cc_library(
     name = "cloud_storage_clients",
     srcs = [
         "abs_client.cc",
+        "abs_client_utils.cc",
         "abs_error.cc",
+        "abs_multipart_state.cc",
         "bucket_name_parts.cc",
         "client_pool.cc",
         "client_probe.cc",
@@ -12,7 +14,9 @@ redpanda_cc_library(
         "credential_manager.cc",
         "multipart_upload.cc",
         "s3_client.cc",
+        "s3_client_utils.cc",
         "s3_error.cc",
+        "s3_multipart_state.cc",
         "upstream.cc",
         "upstream_registry.cc",
         "util.cc",
@@ -20,7 +24,9 @@ redpanda_cc_library(
     ],
     hdrs = [
         "abs_client.h",
+        "abs_client_utils.h",
         "abs_error.h",
+        "abs_multipart_state.h",
         "bucket_name_parts.h",
         "client.h",
         "client_pool.h",
@@ -30,7 +36,9 @@ redpanda_cc_library(
         "logger.h",
         "multipart_upload.h",
         "s3_client.h",
+        "s3_client_utils.h",
         "s3_error.h",
+        "s3_multipart_state.h",
         "types.h",
         "upstream.h",
         "upstream_key.h",
